Code drop of v8mips to Android
This is just a flat drop of the modified and new files in v8mips.
The method used to drop these original files in froyo has no history,
so we are doing the same thing with this drop from v8mips.

This code on v8mips repository will be tagged froyo-drop-1.

Note that the v8mips port is not complete, and there are known bugs.
Specifically a GC bug has required that we disable compacting GC,
the browser will eventually run out of memory and crash.