Build v8 in mips32 mode even for mips16 target.
Building a mixed-mode version of v8 (mips16 C files, with exceptions for
specific files that used mips32 instructions, like ll/sc, and JIT code
in mips32)) yielded a version that ran terribly slow, v8 benchmark score
of 63, Sunspider over 30,000 on 500 MHz 24K.

This change forces all of v8 to build in mips32 mode, even on a mips16
target. Benchmark score back to 129, 11,700.

The total size of libwebcore.so is actually a little bit smaller with
this change, for reasons i do not understand..

Change-Id: I9fe2b462cedfd517c2bfa6b21a313c31bc975919