Add workaround for the 24K cores with the E156 problem. Some 34k cores also have the same problem.
This workaround prevents the jalr-jal optimization if we are hitting
the 24K core E156 problem. Since the excluded space is only 64KB out
of 256MB (0.02%), we are appling the workaround for all cores. That
way we don't have to identify the core type at run time.

Also changing version number to 3.2.3.6.

Change-Id: I8ec24daf6cf405f1d759813257c2e7eba050b071
Signed-off-by: Douglas Leung <douglas@mips.com>