[MIPS] Fix ZeroXBenchmark app crash on v8.
Fix a bug introduced with isolates, where an API LowMemoryNotification()
call made before v8 is initialized will NULL-deref. It is not clear that
it is considered legal to call this API before init, but WebKit currently
does so, and the older (pre-isolates) v8 did not have this problem.

Note that this is an app with a webview which is using WebKit, this is
not a browser issue.

Change v8 version to 3.6.2.5.

Change-Id: I91c4c505816caea1792caf628b0dbfcbd62adfee